requirements.txt
streamlit
matplotlib
plotly
Pillow